LayoutTests/imported/w3c/web-platform-tests/resource-timing/buffer-full-when-populate-entries-expected.txt


FAIL Test that a buffer full event does not bubble and that resourcetimingbufferfull is called only once per overflow assert_equals: Number of entries does not match the expected value. expected 2 but got 0
